hopefully increase the memory limit...

+++ solver.h
@@ -94,29 +94,29 @@ public:
 	}
 	void timedout(){ timeout = true; }
 
-	void solve_ab(const Board & board, double time, int mdepth = 1000);
-	void solve_scout(const Board & board, double time, int mdepth = 1000);
-	void solve_pns(const Board & board, double time, int memlimit);
-	void solve_pnsab(const Board & board, double time, int memlimit);
-	void solve_dfpnsab(const Board & board, double time, int memlimit);
+	void solve_ab     (const Board & board, double time, int mdepth = 1000);
+	void solve_scout  (const Board & board, double time, int mdepth = 1000);
+	void solve_pns    (const Board & board, double time, uint64_t memlimit);
+	void solve_pnsab  (const Board & board, double time, uint64_t memlimit);
+	void solve_dfpnsab(const Board & board, double time, uint64_t memlimit);
 
 //protected:
 
 //used for alpha-beta solvers
 //return -2 for loss, -1,1 for tie, 0 for unknown, 2 for win, all from toplay's perspective
-	int run_negamax(const Board & board, const int depth, int alpha, int beta);  //plain negamax
+	int run_negamax  (const Board & board, const int depth, int alpha, int beta);  //plain negamax
 	int run_negascout(const Board & board, const int depth, int alpha, int beta);  //plain negascout
 
-	int negamax(const Board & board, const int depth, int alpha, int beta);  //plain negamax
+	int negamax  (const Board & board, const int depth, int alpha, int beta);  //plain negamax
 	int negascout(const Board & board, const int depth, int alpha, int beta);  //plain negascout
 
 //basic proof number search building a tree
-	int run_pns(const Board & board, int ties, int memlimit); //1 = win, 0 = unknown, -1 = loss
-	int run_pnsab(const Board & board, int ties, int memlimit); //1 = win, 0 = unknown, -1 = loss
-	int run_dfpnsab(const Board & board, int ties, int memlimit); //1 = win, 0 = unknown, -1 = loss
+	int run_pns    (const Board & board, int ties, uint64_t memlimit); //1 = win, 0 = unknown, -1 = loss
+	int run_pnsab  (const Board & board, int ties, uint64_t memlimit); //1 = win, 0 = unknown, -1 = loss
+	int run_dfpnsab(const Board & board, int ties, uint64_t memlimit); //1 = win, 0 = unknown, -1 = loss
 
-	bool pns(const Board & board, PNSNode * node, int depth);   //basic proof number search
-	bool pnsab(const Board & board, PNSNode * node, int depth); //use a tiny negamax search as a pn,dn heuristic
+	bool pns    (const Board & board, PNSNode * node, int depth);   //basic proof number search
+	bool pnsab  (const Board & board, PNSNode * node, int depth); //use a tiny negamax search as a pn,dn heuristic
 	bool dfpnsab(const Board & board, PNSNode * node, int depth, uint32_t tp, uint32_t td);
 
 //update the phi and delta for the node
