+++ solverpns.cpp
@@ -8,7 +8,7 @@
  * L W L  L   from the perspective of toplay
  * U W LT U
  */
-void Solver::solve_pns(const Board & board, double time, int memlimit){
+void Solver::solve_pns(const Board & board, double time, uint64_t memlimit){
 	reset();
 
 	if(board.won() >= 0){
@@ -42,7 +42,7 @@ void Solver::solve_pns(const Board & board, double time, int memlimit){
 	fprintf(stderr, "Finished in %d msec\n", time_msec() - starttime);
 }
 
-int Solver::run_pns(const Board & board, int ties, int memlimit){ //1 = win, 0 = unknown, -1 = loss
+int Solver::run_pns(const Board & board, int ties, uint64_t memlimit){ //1 = win, 0 = unknown, -1 = loss
 	assignties = ties;
 
 	if(root) delete root;
